Removed decimal_utils.h
