+++ storage/ndb/rest-server/data-access-rondb/src/db-operations/pk/pkr-operation.cpp
@@ -25,7 +25,6 @@
 #include "src/db-operations/pk/pkr-request.hpp"
 #include "src/db-operations/pk/pkr-response.hpp"
 #include "src/db-operations/pk/common.hpp"
-#include "src/rondb-lib/decimal_utils.hpp"
 #include "src/error-strs.h"
 #include "src/logger.hpp"
 #include "src/rdrs-const.h"
